# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Task Management API
  version: 1.0.0
  description: REST API for managing users, projects, tasks, and labels.
servers:
  - url: http://localhost:3000/api
tags:
  - name: Auth
  - name: Projects
  - name: Tasks
  - name: Labels
paths:
  /auth/signup:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupInput'
            example:
              username: owner
              email: owner@example.com
              password: Password123!
              role: USER
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: 4
                username: newuser
                email: newuser@example.com
                role: USER
                created_at: '2026-04-27T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in and receive a JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
            example:
              email: owner@example.com
              password: Password123!
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
              example:
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example.jwt
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /projects:
    get:
      tags: [Projects]
      security:
        - bearerAuth: []
      summary: List accessible projects
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
              example:
                - id: 1
                  user_id: 2
                  name: Database Final Project
                  description: Complete the final project implementation and deployment deliverables.
                  status: IN_PROGRESS
                  due_date: '2026-04-28T23:59:00.000Z'
                  created_at: '2026-04-27T01:00:00.000Z'
                - id: 2
                  user_id: 2
                  name: Sprint Cleanup
                  description: Wrap up follow-up tasks and documentation.
                  status: PLANNING
                  due_date: '2026-05-03T17:00:00.000Z'
                  created_at: '2026-04-27T01:05:00.000Z'
        '401':
          $ref: '#/components/responses/AuthRequired'
    post:
      tags: [Projects]
      security:
        - bearerAuth: []
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectInput'
            example:
              name: Database Final Project
              description: Complete REST API design document
              status: IN_PROGRESS
              due_date: '2026-04-28T23:59:00.000Z'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
              example:
                id: 3
                user_id: 2
                name: API Polish
                description: Finish the final review pass.
                status: PLANNING
                due_date: '2026-05-05T17:00:00.000Z'
                created_at: '2026-04-27T12:15:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthRequired'
  /projects/{id}:
    get:
      tags: [Projects]
      security:
        - bearerAuth: []
      summary: Get a single project
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Project found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
              example:
                id: 1
                user_id: 2
                name: Database Final Project
                description: Complete the final project implementation and deployment deliverables.
                status: IN_PROGRESS
                due_date: '2026-04-28T23:59:00.000Z'
                created_at: '2026-04-27T01:00:00.000Z'
        '400':
          $ref: '#/components/responses/InvalidId'
        '401':
          $ref: '#/components/responses/AuthRequired'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      tags: [Projects]
      security:
        - bearerAuth: []
      summary: Update a project
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdateInput'
            example:
              name: Updated Project Name
              status: COMPLETED
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
              example:
                id: 1
                user_id: 2
                name: Updated Project Name
                description: Complete the final project implementation and deployment deliverables.
                status: COMPLETED
                due_date: '2026-04-28T23:59:00.000Z'
                created_at: '2026-04-27T01:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthRequired'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      tags: [Projects]
      security:
        - bearerAuth: []
      summary: Delete a project
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Project deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteMessage'
              example:
                id: 3
                message: Project deleted successfully
        '400':
          $ref: '#/components/responses/InvalidId'
        '401':
          $ref: '#/components/responses/AuthRequired'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /tasks:
    get:
      tags: [Tasks]
      security:
        - bearerAuth: []
      summary: List accessible tasks
      parameters:
        - in: query
          name: project_id
          schema:
            type: integer
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - in: query
          name: priority
          schema:
            $ref: '#/components/schemas/TaskPriority'
      responses:
        '200':
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
              example:
                - id: 1
                  project_id: 1
                  assigned_user_id: 2
                  title: Build Prisma schema
                  description: Translate the ER diagram into a working Prisma schema.
                  priority: HIGH
                  status: IN_PROGRESS
                  due_date: '2026-04-27T17:00:00.000Z'
                  created_at: '2026-04-27T01:10:00.000Z'
                  labels:
                    - id: 2
                      user_id: 2
                      name: docs
                      color: '#2563eb'
                      created_at: '2026-04-27T01:08:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthRequired'
    post:
      tags: [Tasks]
      security:
        - bearerAuth: []
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskInput'
            example:
              project_id: 1
              assigned_user_id: 2
              title: Build Prisma schema
              description: Translate the ER diagram into Prisma models
              priority: HIGH
              status: IN_PROGRESS
              due_date: '2026-04-27T17:00:00.000Z'
              label_ids: [2]
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              example:
                id: 4
                project_id: 1
                assigned_user_id: 2
                title: Review seed data
                description: Confirm the sample records are usable in Swagger.
                priority: HIGH
                status: TODO
                due_date: '2026-05-01T12:00:00.000Z'
                created_at: '2026-04-27T12:20:00.000Z'
                labels:
                  - id: 1
                    user_id: 2
                    name: bug
                    color: '#ef4444'
                    created_at: '2026-04-27T01:07:00.000Z'
                  - id: 2
                    user_id: 2
                    name: docs
                    color: '#2563eb'
                    created_at: '2026-04-27T01:08:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthRequired'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /tasks/{id}:
    get:
      tags: [Tasks]
      security:
        - bearerAuth: []
      summary: Get a single task
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Task found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              example:
                id: 1
                project_id: 1
                assigned_user_id: 2
                title: Build Prisma schema
                description: Translate the ER diagram into a working Prisma schema.
                priority: HIGH
                status: IN_PROGRESS
                due_date: '2026-04-27T17:00:00.000Z'
                created_at: '2026-04-27T01:10:00.000Z'
                labels:
                  - id: 2
                    user_id: 2
                    name: docs
                    color: '#2563eb'
                    created_at: '2026-04-27T01:08:00.000Z'
        '400':
          $ref: '#/components/responses/InvalidId'
        '401':
          $ref: '#/components/responses/AuthRequired'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      tags: [Tasks]
      security:
        - bearerAuth: []
      summary: Update a task
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateInput'
            example:
              title: Write Swagger documentation
              priority: URGENT
              status: DONE
              label_ids: [1, 3]
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              example:
                id: 1
                project_id: 1
                assigned_user_id: 2
                title: Write Swagger documentation
                description: Translate the ER diagram into a working Prisma schema.
                priority: URGENT
                status: DONE
                due_date: '2026-04-27T17:00:00.000Z'
                created_at: '2026-04-27T01:10:00.000Z'
                labels:
                  - id: 1
                    user_id: 2
                    name: bug
                    color: '#ef4444'
                    created_at: '2026-04-27T01:07:00.000Z'
                  - id: 3
                    user_id: 2
                    name: urgent
                    color: '#f97316'
                    created_at: '2026-04-27T01:09:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthRequired'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      tags: [Tasks]
      security:
        - bearerAuth: []
      summary: Delete a task
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Task deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteMessage'
              example:
                id: 4
                message: Task deleted successfully
        '400':
          $ref: '#/components/responses/InvalidId'
        '401':
          $ref: '#/components/responses/AuthRequired'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /labels:
    get:
      tags: [Labels]
      security:
        - bearerAuth: []
      summary: List accessible labels
      responses:
        '200':
          description: Label list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Label'
              example:
                - id: 1
                  user_id: 2
                  name: bug
                  color: '#ef4444'
                  created_at: '2026-04-27T01:07:00.000Z'
                - id: 2
                  user_id: 2
                  name: docs
                  color: '#2563eb'
                  created_at: '2026-04-27T01:08:00.000Z'
        '401':
          $ref: '#/components/responses/AuthRequired'
    post:
      tags: [Labels]
      security:
        - bearerAuth: []
      summary: Create a label
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabelInput'
            example:
              name: urgent
              color: '#f97316'
      responses:
        '201':
          description: Label created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Label'
              example:
                id: 4
                user_id: 2
                name: frontend
                color: '#22c55e'
                created_at: '2026-04-27T12:30:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthRequired'
        '409':
          $ref: '#/components/responses/ConflictError'
  /labels/{id}:
    get:
      tags: [Labels]
      security:
        - bearerAuth: []
      summary: Get a single label
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Label found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Label'
              example:
                id: 1
                user_id: 2
                name: bug
                color: '#ef4444'
                created_at: '2026-04-27T01:07:00.000Z'
        '400':
          $ref: '#/components/responses/InvalidId'
        '401':
          $ref: '#/components/responses/AuthRequired'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      tags: [Labels]
      security:
        - bearerAuth: []
      summary: Update a label
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabelUpdateInput'
            example:
              name: docs
              color: '#2563eb'
      responses:
        '200':
          description: Label updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Label'
              example:
                id: 1
                user_id: 2
                name: docs
                color: '#2563eb'
                created_at: '2026-04-27T01:07:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthRequired'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
    delete:
      tags: [Labels]
      security:
        - bearerAuth: []
      summary: Delete a label
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Label deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteMessage'
              example:
                id: 4
                message: Label deleted successfully
        '400':
          $ref: '#/components/responses/InvalidId'
        '401':
          $ref: '#/components/responses/AuthRequired'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ResourceId:
      in: path
      name: id
      required: true
      schema:
        type: integer
        minimum: 1
  schemas:
    SignupInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
        created_at:
          type: string
          format: date-time
    ProjectStatus:
      type: string
      enum: [PLANNING, IN_PROGRESS, ON_HOLD, COMPLETED]
    TaskStatus:
      type: string
      enum: [TODO, IN_PROGRESS, DONE, BLOCKED]
    TaskPriority:
      type: string
      enum: [LOW, MEDIUM, HIGH, URGENT]
    ProjectInput:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/ProjectStatus'
        due_date:
          type: string
          format: date-time
    ProjectUpdateInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/ProjectStatus'
        due_date:
          type: string
          format: date-time
    Project:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/ProjectStatus'
        due_date:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    LabelInput:
      type: object
      required: [name, color]
      properties:
        name:
          type: string
        color:
          type: string
    LabelUpdateInput:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
    Label:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        color:
          type: string
        created_at:
          type: string
          format: date-time
    TaskInput:
      type: object
      required: [project_id, title, description]
      properties:
        project_id:
          type: integer
        assigned_user_id:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
        priority:
          $ref: '#/components/schemas/TaskPriority'
        status:
          $ref: '#/components/schemas/TaskStatus'
        due_date:
          type: string
          format: date-time
          nullable: true
        label_ids:
          type: array
          items:
            type: integer
    TaskUpdateInput:
      type: object
      properties:
        project_id:
          type: integer
        assigned_user_id:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
        priority:
          $ref: '#/components/schemas/TaskPriority'
        status:
          $ref: '#/components/schemas/TaskStatus'
        due_date:
          type: string
          format: date-time
          nullable: true
        label_ids:
          type: array
          items:
            type: integer
    Task:
      type: object
      properties:
        id:
          type: integer
        project_id:
          type: integer
        assigned_user_id:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
        priority:
          $ref: '#/components/schemas/TaskPriority'
        status:
          $ref: '#/components/schemas/TaskStatus'
        due_date:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label'
    DeleteMessage:
      type: object
      properties:
        id:
          type: integer
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            type: string
  responses:
    AuthRequired:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Not authenticated. Please provide a valid token.
    UnauthorizedError:
      description: Invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Invalid credentials
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'Forbidden: insufficient permission'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Resource not found
    ConflictError:
      description: Duplicate or conflicting data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Email has already been used
    InvalidId:
      description: Invalid resource ID
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            errors:
              - ID must be a positive integer
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            errors:
              - Name is required